requirements.txt
albumentations==1.3.0
apex==0.9.10dev
einops==0.4.1
einops_exts==0.0.4
face_recognition==1.3.0
h5py==3.1.0
json_tricks==3.16.1
librosa==0.8.0
matplotlib==3.3.4
mmcv==2.0.0
mmcv_full==1.3.17
mmdet==2.25.1
mmtrack==0.14.0
model_archiver==1.0.3
moviepy==1.0.3
munkres==1.1.4
numpy==1.19.5
onnx==1.13.1
onnxruntime==1.14.1
opencv_python==4.7.0.72
pavi==0.0.1
Pillow==9.5.0
psutil==5.9.4
pyrender==0.1.45
pytest==7.3.1
pytorch_sphinx_theme==0.0.19
PyYAML==6.0
requests==2.27.1
scipy==1.5.4
seaborn==0.12.2
setuptools==58.0.4
smplx==0.1.28
soundfile==0.12.1
spacepy==0.4.1
timm==0.6.12
titlecase==2.4
torch==1.8.0+cu111
torchvision==0.9.0+cu111
tensorboard
tqdm==4.64.1
trimesh==3.21.5
ts==0.5.1
typing_extensions==4.5.0
xmltodict==0.13.0
xtcocotools==1.13
